--- Empleos/src/main/java/com/example/Model/Categoria.java
@@ -1,39 +1,0 @@
-package com.example.Model;
-
-
-public class Categoria {
-
-	private Integer id;
-	private String nombre;
-	private Integer descripcion;
-
-	public Integer getId() {
-		return id;
-	}
-
-	public void setId(Integer id) {
-		this.id = id;
-	}
-
-	public String getNombre() {
-		return nombre;
-	}
-
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
-
-	public Integer getDescripcion() {
-		return descripcion;
-	}
-
-	public void setDescripcion(Integer descripcion) {
-		this.descripcion = descripcion;
-	}
-
-	@Override
-	public String toString() {
-		return "Categoria [id=" + id + ", nombre=" + nombre + ", descripcion=" + descripcion + "]";
-	}
-
-}
