+++ Empleos/src/main/java/com/example/demo/Service/ICategoriasService.java
@@ -2,7 +2,7 @@ package com.example.demo.Service;
 
 import java.util.List;
 
-import com.example.Model.Categoria;
+import com.example.demo.model.Categoria;
 
 public interface ICategoriasService {
 	void Guardar(Categoria categoria);
