+++ Empleos/src/main/java/com/example/demo/model/Vacantes.java
@@ -1,0 +1,122 @@
+package com.example.demo.model;
+
+import java.util.Date;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import javax.persistence.Transient;
+
+@Entity
+@Table(name = "Vacantes")
+public class Vacantes {
+	@Id
+	@GeneratedValue(strategy=GenerationType.IDENTITY)
+	private Integer id;
+
+	private String nombre;
+	private String descripcion;
+	private Double salario;
+	private Date fecha;
+	private Integer destacado;
+	private String imagen = "No_picture_available.png";
+	private String detalles;
+	private String estatus;
+	@OneToOne
+	@JoinColumn(name="idCategoria")
+	private Categoria categoria;
+
+	public Integer getId() {
+		return id;
+	}
+
+	public void setId(Integer id) {
+		this.id = id;
+	}
+
+	public String getNombre() {
+		return nombre;
+	}
+
+	public void setNombre(String nombre) {
+		this.nombre = nombre;
+	}
+
+	public String getDescripcion() {
+		return descripcion;
+	}
+
+	public void setDescripcion(String descripcion) {
+		this.descripcion = descripcion;
+	}
+
+	public Double getSalario() {
+		return salario;
+	}
+
+	public void setSalario(Double salario) {
+		this.salario = salario;
+	}
+
+	public Date getFecha() {
+		return fecha;
+	}
+
+	public void setFecha(Date fecha) {
+		this.fecha = fecha;
+	}
+
+	public Integer getDestacado() {
+		return destacado;
+	}
+
+	public void setDestacado(Integer destacado) {
+		this.destacado = destacado;
+	}
+
+	public String getImage() {
+		return imagen;
+	}
+
+	public void setImage(String image) {
+		this.imagen = image;
+	}
+
+	public String getDetalles() {
+		return detalles;
+	}
+
+	public void setDetalles(String detalles) {
+		this.detalles = detalles;
+	}
+
+
+
+	public String getEstatus() {
+		return estatus;
+	}
+
+	public void setEstatus(String estatus) {
+		this.estatus = estatus;
+	}
+
+	public Categoria getCategoria() {
+		return categoria;
+	}
+
+	public void setCategoria(Categoria categoria) {
+		this.categoria = categoria;
+	}
+
+	@Override
+	public String toString() {
+		return "Vacantes [id=" + id + ", nombre=" + nombre + ", descripcion=" + descripcion + ", salario=" + salario
+				+ ", fecha=" + fecha + ", destacado=" + destacado + ", imagen=" + imagen + ", detalles=" + detalles
+				+ ", estatus=" + estatus + ", categoria=" + categoria+"]";
+	}
+
+}
