+++ Empleos/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8091
+server.port=8092
 spring.servlet.multipart.enabled=true
 
 ##//////Windows
@@ -7,3 +7,16 @@ spring.servlet.multipart.location=c:/tmp
 #/spring.servlet.multipart.location=/tmp
 #spring.servlet.multipart.max-file-size=2MB
 empleosapp.ruta.images=c:/empleos/img-vacantes/
+
+
+# DATASOURCE (MYSQL 8.0)
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.datasource.url=jdbc:mysql://localhost:3306/empleosdb?useSSL=false&serverTimezone=America/Mexico_City&allowPublicKeyRetrieval=true
+spring.datasource.username=root
+spring.datasource.password=mysql
+#JPA
+spring.jpa.generate-ddl=false
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
+spring.jpa.show-sql=true
+# Table names physically
+spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
