+++ Empleos/src/main/resources/templates/detalles.html
@@ -27,13 +27,13 @@
 
         <div class="row">
           <div class="col-md-3">            
-            <img class="rounded mx-auto d-block" src="images/logo3.png" alt="Generic placeholder image" width="220" height="220">            
+            <img class="rounded mx-auto d-block" th:src="@{/images/{img} (img=${vacantes.image}) }" alt="Generic placeholder image" width="220" height="220">            
           </div>
           <div class="col-md-9">
             <div class="card">
               <h4 class="card-header">Aplicar para <strong th:text="${vacantes.nombre}"></strong></h4>              
               <div class="card-body">
-                <h5 class="card-title"><strong>Categoría: </strong> <span></span></h5>
+                <h5 class="card-title"><strong>Categoría: </strong> <span th:text="${vacantes.categoria.nombre}"></span></h5>
                 <h6 class="card-title"><strong>Publicado: </strong> <span th:text="${vacantes.fecha}"></span></h6>                
                 <h6 class="card-title"><strong>Salario Ofrecido: </strong ><span th:text="${vacantes.salario}"></span></h6>
                 <p class="card-text">La descripción corta de la oferta de trabajo</p>
@@ -41,7 +41,7 @@
                 <div class="card bg-light mb-3" >
                   <div class="card-body">
                     <h5 class="card-title">Detalles de la oferta de trabajo</h5>
-                    <p class="card-text" th:text="${vacantes.descripcion}"></p>
+                    <p class="card-text" th:utext="${vacantes.detalles}"></p>
                   </div>
                 </div>   
                 <a class="btn btn-primary" href="#" role="button" title="Quiero aplicar para esta oferta de trabajo."><i class="fas fa-file" aria-hidden="true"></i> Aplicar</a>
